order history implemented fully

+++ Ecommerce/EcommerceBetaAPI/EcommerceBetaAPI/Dto/ordersDTO.cs
@@ -5,7 +5,7 @@ namespace EcommerceBetaAPI.Dto
 {
     public class ordersDTO
     {
-        public float TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         public int OrderId { get; set; }
         public int UserId { get; set; }
         public DateTime OrderDate { get; set; }
